requirements.txt
click==8.1.3
jsonschema==3.2.0
polling2==0.5.0
pydantic==1.9.1
PyYAML==6.0
requests==2.27.1
flake8==5.0.4
pytest==7.1.2
colorlog==6.6.0
boto3==1.24.51
botocore==1.27.51